refactor to new package rand
