+++ lightstep/rand/num_gen.go
@@ -1,0 +1,7 @@
+package rand
+
+// NumberGenerator defines an interface to generate numbers.
+type NumberGenerator interface {
+	Int63() int64
+	TwoInt63() (int64, int64)
+}
